% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilabels_creator.R
\name{create.multilabeled.dataset}
\alias{create.multilabeled.dataset}
\title{create.multilabeled.dataset}
\usage{
create.multilabeled.dataset(dataset.path, new.dataset.path)
}
\arguments{
\item{dataset.path}{the root folder of the dataset (optional)}

\item{gene.list}{a list of Gene Symbols of TCGA Genes}

\item{mutation.types}{a list of TCGA mutation Types}

\item{mutation.cons_types}{a list of TCGA mutation Consequence Type (optional)}
}
\description{
This dataset require in input the path (<path_from>) of a directory
where each sub-directory is named after mutation and contain the images
of the cases affected by that mutation. This function will copy
the images into a new folder (<path_to>), specifically into the sub-folders
that represent the label (class) of the images.
}
\details{
The labels will have the following structure: mut([0-9]+).
A file called dictionary.csv will be saved in <path_to> and it will
contains all the pairs <mutation, label>

Furthermore, if an image in labelled with several different mutations,
the name of the directory in which it will be saved will have the
following structure: mutX-nutY-...-mutZ (where X,Y,Z are numbers).
}
\examples{
create.multilabeled.dataset(path_from, path_to)

}
