% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{create.mutation.dataset}
\alias{create.mutation.dataset}
\title{create.mutation.dataset}
\usage{
create.mutation.dataset(gene.list, mutation.types,
  mutation.cons_types = "", dataset.path = "")
}
\arguments{
\item{gene.list}{a list of Gene Symbols of TCGA Genes}

\item{mutation.types}{a list of TCGA mutation Types}

\item{mutation.cons_types}{a list of TCGA mutation Consequence Type (optional)}

\item{dataset.path}{the root folder of the dataset (optional)}
}
\description{
It create the dataset containing all the images available in TCGA related
to the specified mutations. The images will be organized in folders based
on the mutation type, the mutation consequence type and the gene on which
the  mutation had occured. These folders will be placed in the <dataset.path>,
if it specified, or in the current directory otherwise.
}
\examples{
genes =      c("MUC16",            "PIK3CA")
types =      c("Small deletion",   "Small insertion")
cons_types = c("inframe_deletion", "") # The cons_types is optional
#              |________________|
#                       |
#                       |-> Gene + Type + Conseguence Type = First Mutation
# The three vectors (or any iterable data structure) have
# to be created in such a way that:
#   - the first mutation to be considered is: "Small deletion on MUC16 that cause an inframe_deletion"
#   - the second mutation to be cosidered is: "Small insertion in PIK3Ca that cause an inframe_insertion"
#   - ...
#
create.mutation.dataset(genes, types, cons_types)

# DEVELOPER HINT: organize the values in a Data Frame in such a way that the first column contains the Gene,
#                 the second column contains type the and the third column contains con_type. Then the fuction
#                 can be called as follows:  create.mutation.dataset(df$genes, df$types, df$cons_types)

}
