% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{download.images}
\alias{download.images}
\title{download.images}
\usage{
download.images(images, dataset.path = "")
}
\arguments{
\item{images}{the list of TCGA UUID of the images that have to be downloaded}

\item{dataset.path}{the folder in which the images has to be saved (optional)}
}
\description{
Download and save the TCGA images specified using the UUID.
Each image will be saved as <Image UUID>.svs
}
\examples{
# Single image
# Given that 53507819-e29d-46fe-9fdc-fddc6d67a912 is the UUID of a TCGA image
download.images('53507819-e29d-46fe-9fdc-fddc6d67a912', '') # save the image in a folder in the workspace

# List of images
images = c('e7b0b531-2fe3-4b31-bd79-dfeec3b316db', '959e61d6-f78e-4040-90a9-1cbd60f65db6')
download.images(images, '/path/to/a/folder/')

}
