% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{download.mutation.images}
\alias{download.mutation.images}
\title{download.mutation.images}
\usage{
download.mutation.images(gene.symbol, mutation.type,
  mutation.cons_type = "", dataset.path = "")
}
\arguments{
\item{gene.symbol}{the Gene Symbol of a TCGA Gene}

\item{mutation.type}{a TCGA mutation Type}

\item{mutation.cons_type}{a TCGA mutation Consequence Type (optional)}

\item{dataset.path}{the root folder of the dataset (optional)}
}
\description{
Download and save the TCGA images related to one specified mutation type and
mutation consequence type occured on a gene. These image will be saved in a folder
named "<gene_name>_<mutation type>_<mutation consequence type>, which will
be placed in <dataset.path>, if it is specified, or in the current
directory otherwise.
}
\examples{
download.mutation.images("AKAP2", "Single base substitution", "intron_variant")

}
