% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{gene.symbol_to_ensembl}
\alias{gene.symbol_to_ensembl}
\title{gene.symbol_to_ensembl}
\usage{
gene.symbol_to_ensembl(symbol)
}
\arguments{
\item{symbol}{the Gene Symbol of a TCGA Gene}
}
\value{
the Gene Ensembl of the gene
}
\description{
Return the TCGA Gene Ensembl (ex. ENSG00000132694) given the TCGA Gene Symbol (ex. ARHGEF11)
}
\examples{

gene.symbol_to_ensembl("ARHGEF11")

}
