% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia.R
\name{get.BodyPartValues}
\alias{get.BodyPartValues}
\title{get.BodyPartValues}
\usage{
get.BodyPartValues(api.key)
}
\arguments{
\item{apy.key}{the unique TCIA Api Key of the user}
}
\value{
the updated list of body part names
}
\description{
Return the updated list of available body part names
}
\examples{

get.BodyPartValues(<TCIA Api Key>)

}
