% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.list_cases}
\alias{get.list_cases}
\title{get.list_cases}
\usage{
get.list_cases(gene.ensembl, mutation.type, mutation.cons_type = "")
}
\arguments{
\item{gene.ensembl}{a Gene Symbol of a gene that occur in TCGA}

\item{mutation.type}{a TCGA mutation Type (related to the mutations)}

\item{mutation.cons_type}{a TCGA Consequence Type (related to the mutations)}
}
\value{
the list of the case.ids which have the mutation of <mutation.type> on the gene <gene.ensembl>
}
\description{
Return the list of the TCGA case.ids which have the mutation of
TCGA Type <mutation_type> on the gene with Gene Ensembl equal
to <gene.ensembl> that couse the TCGA Consequence Type <mutation.cons_type>
}
\examples{
get.list_cases("ENSG00000152217", "Single base substitution", "missense_variant")

}
