% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia.R
\name{get.ModalityValues}
\alias{get.ModalityValues}
\title{get.ModalityValues}
\usage{
get.ModalityValues(api.key)
}
\arguments{
\item{apy.key}{the unique TCIA Api Key of the user}
}
\value{
the updated list of modality values
}
\description{
Return the updated list of available modality values (i.e. the types of images: CT, MR, ...)
}
\examples{

get.ModalityValues(<TCIA Api Key>)

}
