% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia.R
\name{tcia.create.dataset}
\alias{tcia.create.dataset}
\title{tcia.create.dataset}
\usage{
tcia.create.dataset(api.key, body.parts = "", modality.values = "",
  groupby = "", dataset.path = "", unzip = TRUE)
}
\arguments{
\item{body.parts}{a TCIA body part affected by the tumor (to get the available body part use: get.BodyPartValues)}

\item{modality.values}{a TCIA type of images (CT, MR, ...) that has to be downloaded (to get the update list use: get.ModalityValues)}

\item{groupby}{A list of attribute that can be used to organize in images in nested folders (see example)
(Attribure that can be used: "SeriesInstanceUID","StudyInstanceUID", "Modality", "ProtocolName", "SeriesDate",
"SeriesDescription", "BodyPartExamined", "SeriesNumber", "Collection", "Visibility", "ImageCount", "AnnotationsFlag",
"Manufacturer", "ManufacturerModelName", "SoftwareVersions")}

\item{apy.key}{the unique TCIA Api Key of the user}
}
\description{
Download the TCIA images that match the query parameter (body.parts, modality.values), if specified,
it will download all the images otherwise. Furthermore the images will be organized based on the
<groupby> attributes, if  specified, it will save all the images in the main folder otherwise.
}
\examples{

folder = c("Modality", "StudyInstanceUID")
tcia.create.dataset(<TCIA Api Key>, modality.values = "MR", body.parts = "HEAD", groupby=folder, unzip=FALSE)
# In this case all the magnetic resonance magnetic resonance of the heads, will be download but NOT unzipped.
# These zip files will be saved in the folder: <current directory>/MR/<StudyInstanceUID of the images>
#                                                                  ^^- Modality values of the images
#                                              ^^^^^^^^^^^^^^^^^^- given that dataset.path wasn't spacified

}
