% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia.R
\name{tcia.get.Series}
\alias{tcia.get.Series}
\title{tcia.get.Series}
\usage{
tcia.get.Series(api.key, modality.values, body.parts)
}
\arguments{
\item{modality.values}{a TCIA type of images (CT, MR, ...) that has to be downloaded (to get the update list use: get.ModalityValues)}

\item{body.parts}{a TCIA body part affected by the tumor (to get the available body part use: get.BodyPartValues)}

\item{apy.key}{the unique TCIA Api Key of the user}
}
\value{
a list of the TCIA Series
}
\description{
Return the list of the TCIA Series that match the required parameters
}
\examples{

tcia.get.Series(<TCIA Api Key>, modality.values = "MR", body.parts = "HEAD")

}
